Part 7 attempt at code conversion
The Hurt Signal is in the Hurtbox component but it is never called from there.  The Hitbox component can't reach it to create the signal.  This commit is how it would be converted, but will be moving the signal to the other component in order to enable it to be called