P14 Yellow Enemy
Video 14 in series https://www.youtube.com/playlist?list=PL9FzW-m48fn09w6j8NowI_pSBVcsb3V78

Oops, made the base class include private properties instead of protected

OMG I could not for the life of me figure out why the inheritance wasn't working for the Yellow Enemy.  I think I figured out why:  the MoveComponent was pointing to the root node, which has a script on it.  The default is Enemy.cs and with the Yellow Enemy, we made a new class that extends the Enemy.cs.  However, the MoveComponent isn't able to use this new class even though it extends from the one it knows/expects.  When debugging with the Remote view, the MoveComponent for the YellowEnemy had no actor saved.  I move that root node script to a new node called simply Logic and it worked.  I had to move the reference to get the other nodes since they're now siblings and not children.