P16 Pink Enemy States
Video 16 in series https://www.youtube.com/playlist?list=PL9FzW-m48fn09w6j8NowI_pSBVcsb3V78

Had to add another function to call the signal for StateFinished.  Note this was how Ben did this in the last section and that's what I did earlier when I wasn't sure if it was best practice or not - at least he agreed with me!